[PATRO23-123] Split integration tests to two projects extracting test abstraction
