+++ app.props
@@ -3,7 +3,8 @@
 	<PropertyGroup>
 		<!-- Enable raising errors due not clean source code -->
 		<TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-		 <GenerateDocumentationFile>true</GenerateDocumentationFile>
+		<GenerateDocumentationFile>true</GenerateDocumentationFile>
+		<EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
 		
 		<!-- Code analysis ruleset configuration -->
 		<CodeAnalysisRuleset>$(MSBuildThisFileDirectory)Intive.Patronage2023.ruleset</CodeAnalysisRuleset>
