+++ tests/integration/Intive.Patronage2023.Modules.Example.Application.IntegrationTests/Usings.cs
@@ -1,1 +1,3 @@
+global using Intive.Patronage2023.Shared.IntegrationTests;
+
 global using Xunit;
