+++ tests/integration/Intive.Patronage2023.Shared.IntegrationTests/Usings.cs
@@ -1,0 +1,1 @@
+global using Xunit;
